lora-gateway-main/requirements.txt
blinker==1.8.2
click==8.1.7
dnspython==2.6.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-PyMongo==2.3.0
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mongoengine==0.29.1
pymongo==4.10.1
pyserial==3.5
python-dotenv==1.0.1
Werkzeug==3.0.6
zipp==3.20.2
flask-socketio==5.3.6
